requirements.txt
viam-sdk
numpy
torch
torchvision
pillow
